+++ src/lang/builtins_regexp.ml
@@ -70,7 +70,15 @@ module RegExp = Value.MkAbstract (struct
       (String.concat ""
          (List.sort Stdlib.compare (List.map string_of_regexp_flag flags)))
 
-  let to_json ~compact:_ ~json5:_ r = Utils.quote_string (descr r)
+  let to_json _ =
+    raise
+      Runtime_error.(
+        Runtime_error
+          {
+            kind = "json";
+            msg = "Regexp cannot be represented as json";
+            pos = [];
+          })
 
   let compare r r' =
     Stdlib.compare