+++ src/tools/json_base.ml
@@ -1,0 +1,8 @@
+type t =
+  [ `Assoc of (string * t) list
+  | `Tuple of t list
+  | `String of string
+  | `Bool of bool
+  | `Float of float
+  | `Int of int
+  | `Null ]